feat: add sponsor type to sponsors and update api
The database migration file and lock file were updated to reflect schema changes.
SponsorType enum was introduced to the `Sponsors` model in the prisma schema.
Correspondingly, the `getAll` method in the `SponsorsService` was modified to `getAllFromType`,
taking SponsorType as a parameter.
The `getV1` method in the `SponsorsController` was also adjusted to accommodate this change.