feat: add countdown and registration start date validation
Implemented a countdown component that displays the time remaining to event commencement.
Add the countdown component into the sign-up form when the current time is earlier than the defined
registration start date.
Updated the user's service to set the nickname as a full name consisting of the first and last name.