feat: add team-ranking feature and points calculation
Introduced a new feature that sorts and ranks teams based on their points scored in the hackathon.
To facilitate this, the schema has been updated to include team points, and the teams service now
calculates and includes the total points in its response.
Various UI changes were made to accommodate this update, including the addition of a ranking page.